test: move github-issue-70 test to cascades/cascade-remove-one-to-many
